Compiled for 64 bit arch

+++ quickdialog/QDateTimeInlineElement.m
@@ -95,7 +95,7 @@
     NSInteger seconds = ti % 60;
     NSInteger minutes = (ti / 60) % 60;
     NSInteger hours = (ti / 3600);
-    return [NSString stringWithFormat:@"%02i:%02i:%02i", hours, minutes, seconds];
+    return [NSString stringWithFormat:@"%02li:%02li:%02li", (long)hours, (long)minutes, (long)seconds];
 }
 
 - (void)fetchValueIntoObject:(id)obj {
